requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.13.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.0.3
gdown==5.2.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.68.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
PyMuPDF==1.24.11
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing-extensions==4.13.0
uritemplate==4.1.1
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
